before minseo
