+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "html.format.enable": false
+}
